Make Search generic
