+++ src/jsMain/kotlin/io/beatmaps/user/alerts.kt
@@ -7,7 +7,7 @@ import external.generateConfig
 import io.beatmaps.api.AlertUpdate
 import io.beatmaps.api.UserAlert
 import io.beatmaps.common.Config
-import io.beatmaps.index.coloredCard
+import io.beatmaps.shared.coloredCard
 import io.beatmaps.setPageTitle
 import io.beatmaps.util.textToContent
 import kotlinx.html.js.onClickFunction
